Move search routes to their own router
